+++ OneMore/Helpers/Office/PowerPoint.cs
@@ -8,6 +8,7 @@ namespace River.OneMoreAddIn.Helpers.Office
 	using Microsoft.Office.Interop.PowerPoint;
 	using System;
 	using System.IO;
+	using System.Runtime.InteropServices;
 
 
 	internal class PowerPoint : IDisposable
@@ -29,6 +30,8 @@ namespace River.OneMoreAddIn.Helpers.Office
 			if (!disposed)
 			{
 				power.Quit();
+				Marshal.ReleaseComObject(power);
+
 				power = null;
 				disposed = true;
 			}
